tests/requirements/dj41_cms311.txt
-r base.txt

Django>=4.1,<4.2
django-cms>=3.11.1,<4
